+++ src/renderer.js
@@ -2,3 +2,5 @@ export { default as createImageRenderer } from './image-renderer';
 export { default as createMatrixRenderer } from './matrix-renderer';
 export { default as createRepresentativeRenderer } from './representative-renderer';
 export { default as createSvgRenderer } from './svg-renderer';
+export { default as createD3Renderer } from './d3-renderer';
+export { default as createVegaLiteRenderer } from './vega-lite-renderer';